+++ pacai/core/game.py
@@ -9,7 +9,6 @@ import traceback
 import sys
 
 from pacai.util import util
-from pacai.util import interrupts
 
 class Directions:
     NORTH = 'North'
@@ -598,14 +597,14 @@ class Game:
             self.mute()
             if self.catchExceptions:
                 try:
-                    timed_func = interrupts.TimeoutFunction(agent.registerInitialState,
+                    timed_func = util.TimeoutFunction(agent.registerInitialState,
                             int(self.rules.getMaxStartupTime(i)))
                     try:
                         start_time = time.time()
                         timed_func(self.state.deepCopy())
                         time_taken = time.time() - start_time
                         self.totalAgentTimes[i] += time_taken
-                    except interrupts.TimeoutFunctionException:
+                    except util.TimeoutFunctionException:
                         logging.warning('Agent %d ran out of time on startup!' % i)
                         self.unmute()
                         self.agentTimeout = True
@@ -632,12 +631,12 @@ class Game:
             self.mute()
             if self.catchExceptions:
                 try:
-                    timed_func = interrupts.TimeoutFunction(agent.observationFunction,
+                    timed_func = util.TimeoutFunction(agent.observationFunction,
                             int(self.rules.getMoveTimeout(agentIndex)))
                     try:
                         start_time = time.time()
                         observation = timed_func(self.state.deepCopy())
-                    except interrupts.TimeoutFunctionException:
+                    except util.TimeoutFunctionException:
                         skip_action = True
                     move_time += time.time() - start_time
                     self.unmute()
@@ -654,14 +653,14 @@ class Game:
             self.mute()
             if self.catchExceptions:
                 try:
-                    timed_func = interrupts.TimeoutFunction(agent.getAction,
+                    timed_func = util.TimeoutFunction(agent.getAction,
                             int(self.rules.getMoveTimeout(agentIndex)) - int(move_time))
                     try:
                         start_time = time.time()
                         if skip_action:
-                            raise interrupts.TimeoutFunctionException()
+                            raise util.TimeoutFunctionException()
                         action = timed_func(observation)
-                    except interrupts.TimeoutFunctionException:
+                    except util.TimeoutFunctionException:
                         logging.warning('Agent %d timed out on a single move!' % agentIndex)
                         self.agentTimeout = True
                         self.unmute()
